Add all todo functionalities to CRA

+++ packages/react-scripts/scripts/init.js
@@ -39,9 +39,13 @@ module.exports = function(
 
   // Setup the script rules
   appPackage.scripts = {
-    start: 'react-scripts start',
+    start: 'webpack-dashboard -c cyan -- react-scripts start',
     build: 'react-scripts build',
-    test: 'react-scripts test --env=jsdom',
+    pretest: 'npm run test:lint',
+    test: 'npm run test:unit',
+    'test:unit': 'react-scripts test:unit',
+    'test:watch': 'react-scripts test:watch',
+    'test:lint': './node_modules/.bin/eslint . --color; exit 0',
     eject: 'react-scripts eject',
   };
 
@@ -91,17 +95,38 @@ module.exports = function(
     }
   );
 
+  // Rename npmrc after the fact to prevent npm from deleting it
+  fs.move(
+    path.join(appPath, 'npmrc'),
+    path.join(appPath, '.npmrc'),
+    [],
+    err => {
+      if (err) {
+        // Overwrite if there's already a `.npmrc` file there
+        if (err.code === 'EEXIST') {
+          const data = fs.readFileSync(path.join(appPath, 'npmrc'));
+          fs.writeFileSync(path.join(appPath, '.npmrc'), data);
+          fs.unlinkSync(path.join(appPath, 'npmrc'));
+        } else {
+          throw err;
+        }
+      }
+    }
+  );
+
   let command;
-  let args;
+  let save;
+  let saveDev;
 
   if (useYarn) {
     command = 'yarnpkg';
-    args = ['add'];
+    save = ['add'];
   } else {
     command = 'npm';
-    args = ['install', '--save', verbose && '--verbose'].filter(e => e);
+    save = ['install', '--save', verbose && '--verbose'].filter(e => e);
+    saveDev = ['install', '--save-dev', verbose && '--verbose'].filter(e => e);
   }
-  args.push('react', 'react-dom');
+  save.push('react', 'react-dom');
 
   // Install additional template dependencies, if present
   const templateDependenciesPath = path.join(
@@ -110,11 +135,32 @@ module.exports = function(
   );
   if (fs.existsSync(templateDependenciesPath)) {
     const templateDependencies = require(templateDependenciesPath).dependencies;
-    args = args.concat(
+    save = save.concat(
       Object.keys(templateDependencies).map(key => {
         return `${key}@${templateDependencies[key]}`;
       })
     );
+    const templateDevDependencies = require(templateDependenciesPath).devDependencies;
+    saveDev = saveDev.concat(
+      Object.keys(templateDevDependencies).map(key => {
+        return `${key}@${templateDevDependencies[key]}`;
+      })
+    );
+
+    console.log(`Installing templated packages using ${command}...`);
+    console.log();
+
+    let proc = spawn.sync(command, save, { stdio: 'inherit' });
+    if (proc.status !== 0) {
+      console.error(`\`${command} ${save.join(' ')}\` failed`);
+      return;
+    }
+    proc = spawn.sync(command, saveDev, { stdio: 'inherit' });
+    if (proc.status !== 0) {
+      console.error(`\`${command} ${saveDev.join(' ')}\` failed`);
+      return;
+    }
+
     fs.unlinkSync(templateDependenciesPath);
   }
 
@@ -125,9 +171,14 @@ module.exports = function(
     console.log(`Installing react and react-dom using ${command}...`);
     console.log();
 
-    const proc = spawn.sync(command, args, { stdio: 'inherit' });
+    let proc = spawn.sync(command, save, { stdio: 'inherit' });
     if (proc.status !== 0) {
-      console.error(`\`${command} ${args.join(' ')}\` failed`);
+      console.error(`\`${command} ${save.join(' ')}\` failed`);
+      return;
+    }
+    proc = spawn.sync(command, saveDev, { stdio: 'inherit' });
+    if (proc.status !== 0) {
+      console.error(`\`${command} ${saveDev.join(' ')}\` failed`);
       return;
     }
   }
@@ -146,7 +197,9 @@ module.exports = function(
   const displayedCommand = useYarn ? 'yarn' : 'npm';
 
   console.log();
-  console.log(`Success! Created ${appName} at ${appPath}`);
+  console.log(`Success! Created ${appName} at ${appPath} with ` +
+    chalk.cyan(`cc-react-scripts`)
+  );
   console.log('Inside that directory, you can run several commands:');
   console.log();
   console.log(chalk.cyan(`  ${displayedCommand} start`));
